+++ ui/src/app/core/accessories/types/humidity-sensor/humidity-sensor.component.ts
@@ -1,5 +1,5 @@
-import { NgClass } from '@angular/common'
 import { Component, Input } from '@angular/core'
+import { TranslatePipe } from '@ngx-translate/core'
 
 import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
 
@@ -9,7 +9,7 @@ import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
   styleUrls: ['./humidity-sensor.component.scss'],
   standalone: true,
   imports: [
-    NgClass,
+    TranslatePipe,
   ],
 })
 export class HumiditySensorComponent {