+++ ui/src/app/core/settings.service.ts
@@ -24,8 +24,11 @@ export class SettingsService {
   public env: EnvInterface = {} as EnvInterface
   public host: string
   public proxyHost: string
+  public webroot: string
+  public originalWebroot: string
   public formAuth = true
   public sessionTimeout = 28800
+  public sessionTimeoutInactivityBased = false
   public uiVersion: string
   public theme: string
   public lightingMode: 'auto' | 'light' | 'dark'
@@ -64,8 +67,11 @@ export class SettingsService {
     const data = await firstValueFrom(this.$api.get('/auth/settings')) as AppSettingsInterface
     this.formAuth = data.formAuth
     this.sessionTimeout = data.sessionTimeout
+    this.sessionTimeoutInactivityBased = data.sessionTimeoutInactivityBased
     this.env = data.env
     this.host = data.host
+    this.webroot = data.webroot
+    this.originalWebroot = data.originalWebroot
     this.proxyHost = data.proxyHost
     this.lightingMode = data.lightingMode
     this.wallpaper = data.wallpaper