+++ ui/src/app/modules/settings/remove-all-accessories/remove-all-accessories.component.ts
@@ -38,7 +38,9 @@ export class RemoveAllAccessoriesComponent implements OnInit {
           this.$translate.instant('toast.title_success'),
         )
         this.$activeModal.close()
-        this.$router.navigate(['/restart'], { queryParams: { restarting: true } })
+        void this.$router.navigate(['/restart'], {
+          queryParams: { restarting: true },
+        })
       },
       error: (error) => {
         this.clicked = false